modified src/base_classes.py, src/models.py
